EX-16 hecho y fix en EX-12 (De Float a Double)

+++ Practica_4/EX-11.hs
@@ -1,10 +1,10 @@
-eAprox :: Integer -> Float
+eAprox :: Integer -> Double
 eAprox 0 = 1
 eAprox n = (1 / (fact (fromIntegral n))) + eAprox (n-1)
 
-fact :: Float -> Float 
+fact :: Double -> Double 
 fact 1 = 1
 fact n = n * fact (n-1) 
 
-e :: Float
+e :: Double
 e = eAprox 10
